+++ src/features/basic/header-calculator-button.ts
@@ -7,6 +7,7 @@ async function onTileReady(tile: PrunTile) {
   createFragmentApp(TileControlsButton, {
     icon: '\uf1ec',
     onClick: () => showBuffer('XIT CALC'),
+    marginTop: 4,
   }).before(tileControls.children[0]);
   return;
 }